Re-structure files
