+++ graph/definition/auth/auth.graphql
@@ -1,0 +1,8 @@
+type LoginUserResponse {
+    id: ID!
+    token: String!
+}
+
+extend type Mutation {
+    login(username: String!, password: String!): LoginUserResponse!
+}
