+++ graph/definition/user/user.graphql
@@ -9,11 +9,6 @@ type CreateUserResponse {
     id: ID!
 }
 
-type LoginUserResponse {
-    id: ID!
-    token: String!
-}
-
 type User {
     id: ID!
     name: String!
@@ -32,6 +27,5 @@ input NewUser {
 }
 
 type Mutation {
-    createUser(newUser: NewUser!): CreateUserResponse! @isAdmin
-    login(username: String!, password: String!): LoginUserResponse!
+    user(newUser: NewUser!): CreateUserResponse! @isAdmin
 }
