+++ graph/resolver/schema.directives.go
@@ -1,4 +1,4 @@
-package graph
+package resolver
 
 import (
 	"context"
